third_party/blink/web_tests/virtual/selection-shadow-delegates-focus-disabled/shadow-dom/focus-slide-on-shadow-host-expected.txt

This is a testharness.js-based test.
[FAIL] click on inner div should focus inner focusable (with delegatesFocus = true).
  assert_equals: expected 0 but got 1
Harness: the test ran to completion.
